+++ vscode-extension/src/alsyntaxmodifiers/removeRedundantDataClassificationModifier.ts
@@ -1,0 +1,10 @@
+import { DevToolsExtensionContext } from "../devToolsExtensionContext";
+import { WorkspaceCommandSyntaxModifier } from "./workspaceCommandSyntaxModifier";
+
+export class RemoveRedundantDataClassificationModifier extends WorkspaceCommandSyntaxModifier {
+
+    constructor(context: DevToolsExtensionContext) {
+        super(context, "Remove Redundant DataClassification", "removeRedundantDataClassification");
+    }
+
+}